requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
numpy==1.26.2
python-dotenv==1.0.0
cryptography==41.0.7
bleach==6.1.0